Refactoring NLP classifiers, making them all exportable, adding/moving unit tests
