+++ tests/collections/nlp/test_nlp_exportables.py
@@ -1,0 +1,61 @@
+# Copyright (c) 2020, NVIDIA CORPORATION.  All rights reserved.
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#     http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+import os
+import tempfile
+
+import pytest
+from omegaconf import DictConfig
+
+from nemo.collections.nlp.modules.common import (
+    BertPretrainingTokenClassifier,
+    SequenceClassifier,
+    SequenceRegression,
+    SequenceTokenClassifier,
+    TokenClassifier,
+)
+
+
+def classifier_export(obj):
+    with tempfile.TemporaryDirectory() as tmpdir:
+        filename = os.path.join(tmpdir, obj.__class__.__name__ + '.onnx')
+        obj.export(output=filename)
+
+
+class TestExportableClassifiers:
+    @pytest.mark.unit
+    def test_token_classifier_export_to_onnx(self):
+        for num_layers in [1, 2, 4]:
+            classifier_export(TokenClassifier(hidden_size=256, num_layers=num_layers, num_classes=16))
+
+    @pytest.mark.unit
+    def test_bert_pretraining_export_to_onnx(self):
+        for num_layers in [1, 2, 4]:
+            classifier_export(TokenClassifier(hidden_size=256, num_layers=num_layers, num_classes=16))
+
+    @pytest.mark.unit
+    def test_sequence_token_classifier_export_to_onnx(self):
+        for num_layers in [1, 2, 4]:
+            classifier_export(
+                SequenceTokenClassifier(hidden_size=256, num_slots=8, num_intents=8, num_layers=num_layers)
+            )
+
+    @pytest.mark.unit
+    def test_sequence_classifier_export_to_onnx(self):
+        for num_layers in [1, 2, 4]:
+            classifier_export(SequenceClassifier(hidden_size=256, num_classes=16, num_layers=num_layers))
+
+    @pytest.mark.unit
+    def test_sequence_regression_export_to_onnx(self):
+        for num_layers in [1, 2, 4]:
+            classifier_export(SequenceRegression(hidden_size=256, num_layers=num_layers))
